[FIX] pantalla Jugar_Amigos.kt refactorizada en componentes, con punto de entrada para API (usecase) y fix a errores
