+++ SerpientesYEscalerasReMix/app/src/main/java/com/UNIZAR_30226_2026_10/SerpientesYEscalerasReMix/domain/usecase/CaseFacade.kt
@@ -8,9 +8,11 @@ class CaseFacade(context: Context) {
     // Creación del almacen local
     private val local = LocalStorage(context)
 
-    // Creación de la clase remote relacionada con la API
+    // Creación de la clase remote relacionada con la API TODO
     // private val remoteApi =
 
     // Crear Todos los casos de uso, asignando local y remoteApi segun corresponda
     public val loginRegisterCase = LoginRegisterCase(local)
+
+    public val amigosCase = AmigosCase()
 }
