+++ SerpientesYEscalerasReMix/app/src/main/java/com/UNIZAR_30226_2026_10/SerpientesYEscalerasReMix/ui/components/listaTarjetasJugarAmigos.kt
@@ -1,0 +1,192 @@
+package com.UNIZAR_30226_2026_10.SerpientesYEscalerasReMix.ui.components
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.border
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.interaction.MutableInteractionSource
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.AddCircle
+import androidx.compose.material.icons.filled.Delete
+import androidx.compose.material.icons.filled.Email
+import androidx.compose.material.icons.filled.PlayArrow
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.draw.rotate
+import androidx.compose.ui.draw.shadow
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.unit.dp
+import com.UNIZAR_30226_2026_10.SerpientesYEscalerasReMix.R
+import com.UNIZAR_30226_2026_10.SerpientesYEscalerasReMix.ui.screens.Amigo
+import com.UNIZAR_30226_2026_10.SerpientesYEscalerasReMix.ui.theme.SETextTypes
+import com.UNIZAR_30226_2026_10.SerpientesYEscalerasReMix.ui.theme.color_bg
+import com.UNIZAR_30226_2026_10.SerpientesYEscalerasReMix.ui.theme.color_offline
+import com.UNIZAR_30226_2026_10.SerpientesYEscalerasReMix.ui.theme.color_online
+import com.UNIZAR_30226_2026_10.SerpientesYEscalerasReMix.ui.theme.color_secondary
+import com.UNIZAR_30226_2026_10.SerpientesYEscalerasReMix.ui.theme.color_selected
+import com.UNIZAR_30226_2026_10.SerpientesYEscalerasReMix.ui.theme.color_text
+import com.UNIZAR_30226_2026_10.SerpientesYEscalerasReMix.ui.theme.color_unselected
+
+
+@Composable
+fun ListaAmigos(amigos: List<Amigo>) {
+    // Guarda de forma observable el amigo del cual se estan viendo los detalles
+    var amigoExpandido by remember { mutableStateOf<String?>(null) }
+
+    LazyColumn(
+        modifier = Modifier.fillMaxSize(),
+        verticalArrangement = Arrangement.spacedBy(0.dp)
+    ) {
+        items(amigos) { amigo ->
+            TarjetaAmigo(amigo, amigoExpandido == amigo.nombre, onClick = {
+                amigoExpandido = if (amigoExpandido == amigo.nombre) null
+                else amigo.nombre
+            })
+        }
+    }
+}
+
+@Composable
+fun TarjetaAmigo(amigo: Amigo, expandido: Boolean, onClick: () -> Unit) {
+    // Si te ha invitado, se usa azul claro
+    val fondo = if (amigo.haInvitado) color_selected
+    else if (amigo.estaOnline) color_secondary
+    else color_unselected
+
+    Surface(
+        color = fondo,
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(vertical = 20.dp),
+        shadowElevation = 4.dp
+    ) {
+        Column() {
+            informacionAmigo(amigo, expandido, onClick)
+
+            if (expandido) {
+                desplegableAmigo(amigo)
+            }
+        }
+
+    }
+}
+
+@Composable
+fun informacionAmigo(amigo: Amigo, expandido: Boolean, onClick: () -> Unit) {
+    Row(
+        modifier = Modifier
+            .padding(horizontal = 16.dp, vertical = 12.dp)
+            .clickable(interactionSource = remember { MutableInteractionSource() }, indication = null) { onClick() }
+            .fillMaxWidth(),
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        // Avatar con indicador de estado
+        Box(contentAlignment = Alignment.BottomEnd) {
+            Image(
+                painter = painterResource(id = R.drawable.amigos),
+                contentDescription = null,
+                modifier = Modifier
+                    .size(60.dp)
+                    .shadow(
+                        elevation = 10.dp,
+                        shape = CircleShape,
+                        clip = false
+                    )
+                    .clip(CircleShape)
+                    .border(1.dp, color_bg, CircleShape)
+            )
+            // Círculo de estado (Verde/Rojo)
+            Surface(
+                shape = CircleShape,
+                color = if (amigo.estaOnline) color_online else color_offline,
+                modifier = Modifier.size(20.dp)
+            ) {}
+        }
+
+        Spacer(modifier = Modifier.width(16.dp))
+
+        // Nombre y si ha invitado o no
+        Column(modifier = Modifier.weight(1f)) {
+            Text(
+                text = amigo.nombre,
+                style = SETextTypes.plano,
+                color = color_text
+            )
+            Text(
+                text = amigo.estadoTexto,
+                style = SETextTypes.sombreado,
+                color = color_text.copy(alpha = 0.7f)
+            )
+        }
+
+        // Icono de flecha
+        // rotación de la flecha (en caso de que se haber extendio el item)
+        var rotacion = if (expandido) 90f
+        else 0f
+        Icon(
+            imageVector = Icons.Default.PlayArrow,
+            contentDescription = null,
+            modifier = Modifier
+                .size(16.dp)
+                .rotate(rotacion),
+            tint = color_text
+        )
+    }
+}
+
+@Composable
+fun desplegableAmigo(amigo: Amigo) {
+    Column(
+        modifier = Modifier
+            .fillMaxWidth()
+            .background(color_secondary.copy(alpha = 0.5f)) // Un tono ligeramente distinto
+            .padding(bottom = 8.dp, start = 16.dp, end = 16.dp)
+    ) {
+
+        val colorUnirse = if (amigo.haInvitado) color_text
+        else color_unselected
+
+        val acciones = listOf(
+            Triple(Icons.Default.AddCircle, "Invitar a la partida", color_text),
+            Triple(Icons.Default.PlayArrow, "Unirse a la partida", colorUnirse),
+            Triple(Icons.Default.Email, "Enviar un mensaje", color_text),
+            Triple(Icons.Default.Delete, "Borrar amigo", color_text)
+        )
+
+        acciones.forEach { (icono, texto, color) ->
+            Row(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .clickable { /* Acción aquí */ }
+                    .padding(vertical = 10.dp),
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                Icon(icono, contentDescription = null, tint = color, modifier = Modifier.size(24.dp))
+                Spacer(modifier = Modifier.width(12.dp))
+                Text(texto, style = SETextTypes.plano, color = color)
+            }
+        }
+    }
+}
