+++ SerpientesYEscalerasReMix/app/src/main/java/com/UNIZAR_30226_2026_10/SerpientesYEscalerasReMix/ui/navigation/SENavGraph.kt
@@ -56,7 +56,7 @@ fun NavGraphBuilder.navGraph(SEState: SENavHostController, snackHost: SnackbarHo
             )
         }
     ) {
-        AmigosScreen(SEState)
+        AmigosScreen(SEState, cF)
     }
 
     composable(Destinos.MAZOS) {
